refactor : 클래스 책임 분리
