+++ src/main/java/player/Attacker.java
@@ -1,3 +1,5 @@
+package player;
+
 import java.util.Random;
 
 public class Attacker extends Player{
