+++ src/main/java/player/Midfielder.java
@@ -1,3 +1,5 @@
+package player;
+
 import java.util.Random;
 
 public class Midfielder extends Player{
