+++ src/main/java/player/Player.java
@@ -1,3 +1,5 @@
+package player;
+
 abstract public class Player{
     private String name;
     private int score;
